+++ src/node-lib/curriculum-api-2023/fixtures/subjectPhaseOptions.fixture.ts
@@ -8,11 +8,15 @@ const subjectPhaseOptionsFixture = (): SubjectPhaseOption[] => [
       { title: "Primary", slug: "primary" },
       { title: "Secondary", slug: "secondary" },
     ],
-    examboards: [
+    ks4_options: [
       { title: "AQA", slug: "aqa" },
       { title: "Edexcel", slug: "edexcel" },
     ],
     cycle: "1",
+    keystages: [
+      { title: "KS1", slug: "ks1" },
+      { title: "KS3", slug: "ks3" },
+    ],
   },
   {
     title: "Geography",
@@ -21,8 +25,12 @@ const subjectPhaseOptionsFixture = (): SubjectPhaseOption[] => [
       { title: "Primary", slug: "primary" },
       { title: "Secondary", slug: "secondary" },
     ],
-    examboards: null,
+    ks4_options: null,
     cycle: "1",
+    keystages: [
+      { title: "KS1", slug: "ks1" },
+      { title: "KS3", slug: "ks3" },
+    ],
   },
   {
     title: "History",
@@ -31,11 +39,15 @@ const subjectPhaseOptionsFixture = (): SubjectPhaseOption[] => [
       { title: "Primary", slug: "primary" },
       { title: "Secondary", slug: "secondary" },
     ],
-    examboards: [
+    ks4_options: [
       { title: "AQA", slug: "aqa" },
       { title: "Edexcel", slug: "edexcel" },
     ],
     cycle: "1",
+    keystages: [
+      { title: "KS1", slug: "ks1" },
+      { title: "KS3", slug: "ks3" },
+    ],
   },
   {
     title: "Maths",
@@ -44,15 +56,23 @@ const subjectPhaseOptionsFixture = (): SubjectPhaseOption[] => [
       { title: "Primary", slug: "primary" },
       { title: "Secondary", slug: "secondary" },
     ],
-    examboards: null,
+    ks4_options: null,
     cycle: "1",
+    keystages: [
+      { title: "KS1", slug: "ks1" },
+      { title: "KS3", slug: "ks3" },
+    ],
   },
   {
     title: "Music",
     slug: "music",
     phases: [{ title: "Secondary", slug: "secondary" }],
-    examboards: null,
+    ks4_options: null,
     cycle: "1",
+    keystages: [
+      { title: "KS1", slug: "ks1" },
+      { title: "KS3", slug: "ks3" },
+    ],
   },
   {
     title: "Science",
@@ -61,8 +81,12 @@ const subjectPhaseOptionsFixture = (): SubjectPhaseOption[] => [
       { title: "Primary", slug: "primary" },
       { title: "Secondary", slug: "secondary" },
     ],
-    examboards: null,
+    ks4_options: null,
     cycle: "1",
+    keystages: [
+      { title: "KS1", slug: "ks1" },
+      { title: "KS3", slug: "ks3" },
+    ],
   },
 ];
 export default subjectPhaseOptionsFixture;