+++ src/pages/api/curriculum-downloads/index.ts
@@ -195,9 +195,9 @@ async function getData(opts: {
   const subject = subjectPhaseOptions.subjects.find((subject) => {
     return subject.slug === subjectSlug && subject.state === state;
   }) as SubjectPhasePickerData["subjects"][number] | undefined;
-  const examboard =
-    subject?.examboards?.find(
-      (examboard) => examboard.slug === examboardSlug,
+  const ks4Option =
+    subject?.ks4_options?.find(
+      (ks4_option) => ks4_option.slug === examboardSlug,
     ) ?? null;
 
   const combinedCurriculumData: CombinedCurriculumData = {
@@ -205,7 +205,7 @@ async function getData(opts: {
     ...curriculumOverviewTabData,
     ...curriculumOverviewSanityData,
     ...{ state },
-    examboardTitle: examboard?.title ?? null,
+    examboardTitle: ks4Option?.title ?? null,
   };
 
   return {